YoloV8/requirements.txt
numpy
pandas
scikit-learn
ultralytics
wandb
python-dotenv
